+++ test/unit/tracks/text/text-track-list.test.js
@@ -1,6 +1,6 @@
-import TextTrackList from '../../../src/js/tracks/text-track-list.js';
-import TextTrack from '../../../src/js/tracks/text-track.js';
-import EventTarget from '../../../src/js/event-target.js';
+import TextTrackList from '../../../../src/js/tracks/text/text-track-list.js';
+import TextTrack from '../../../../src/js/tracks/text/text-track.js';
+import EventTarget from '../../../../src/js/event-target.js';
 
 var noop = Function.prototype;
 var genericTracks = [
