+++ test/unit/tracks/text/text-track-settings.test.js
@@ -1,6 +1,6 @@
-import TextTrackSettings from '../../../src/js/tracks/text-track-settings.js';
-import TestHelpers from '../test-helpers.js';
-import * as Events from '../../../src/js/utils/events.js';
+import TextTrackSettings from '../../../../src/js/tracks/text/text-track-settings.js';
+import TestHelpers from '../../test-helpers.js';
+import * as Events from '../../../../src/js/utils/events.js';
 import safeParseTuple from 'safe-json-parse/tuple';
 import window from 'global/window';
